Update SDK version to the latest major release

+++ src/main/java/uk/gov/companieshouse/web/accounts/model/smallfull/CalledUpShareCapitalNotPaid.java
@@ -7,5 +7,5 @@ import lombok.Setter;
 @Setter
 public class CalledUpShareCapitalNotPaid {
 
-    private Integer currentAmount;
+    private Long currentAmount;
 }
