+++ src/main/java/uk/gov/companieshouse/web/accounts/model/smallfull/TangibleAssets.java
@@ -7,6 +7,6 @@ import lombok.Setter;
 @Setter
 public class TangibleAssets {
 
-    private Integer currentAmount;
+    private Long currentAmount;
 
 }
